feat: adding game metrics
