+++ services/game/game-manager/Dockerfile
@@ -9,4 +9,6 @@ RUN npm install
 COPY src ./src
 COPY test ./test
 
+EXPOSE 5050
+
 CMD ["npm", "start"]
