+++ services/game/game-manager/package.json
@@ -4,7 +4,7 @@
 	"main": "src/index.js",
 	"type": "module",
 	"scripts": {
-		"start": "node src/index.js && node src/metrics.js",
+		"start": "node src/index.js",
 		"dev": "nodemon src/index.js",
 		"test": "vitest"
 	},
