+++ services/game/pong-physics/Dockerfile
@@ -9,4 +9,6 @@ RUN npm install
 COPY src ./src
 COPY test ./test
 
+EXPOSE 5052
+
 CMD ["npm", "start"]
