+++ services/game/pongBR-physics/package.json
@@ -10,6 +10,7 @@
   },
   "dependencies": {
     "blessed": "^0.1.81",
+    "fastify": "^5.2.1",
     "blessed-contrib": "^4.11.0",
     "dotenv": "^16.0.3",
     "nats": "^2.29.3",
