+++ services/game/pongBR-physics/src/index.js
@@ -66,3 +66,37 @@ async function start() {
 }
 
 start();
+
+
+// -------------------------------------------------------------------------------
+//                              Metrics server
+
+import Fastify from 'fastify';
+import client from 'prom-client';
+import fs from 'fs';
+
+const app = Fastify({
+	logger: true,
+	https: {
+		key: fs.readFileSync(process.env.SSL_KEY),
+		cert: fs.readFileSync(process.env.SSL_CERT)
+	}
+});
+
+client.collectDefaultMetrics();
+
+app.get('/metrics', async (req, res) => {
+	res.type('text/plain');
+	res.send(await client.register.metrics());
+});
+
+const startF = async () => {
+	try {
+		await app.listen({ port: 5053, host: '0.0.0.0' });
+	} catch (err) {
+		app.log.error(err);
+		process.exit(1);
+	}
+};
+
+startF();
