+++ services/game/user-interface/package.json
@@ -10,6 +10,7 @@
   },
   "dependencies": {
     "dotenv": "^16.0.3",
+    "fastify": "^5.2.1",
     "nats": "^2.29.3",
     "prom-client": "^15.1.3",
     "protobufjs": "^7.5.0",
